Move deprecated wrappers to swanlab/deprecated subpackage
